# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AgriSmart Agriculture Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /equipment/listings:
    get:
      operationId: getEquipmentListings
      tags: [equipment]
      summary: Get available equipment for rent
      parameters:
        - name: location
          in: query
          required: false
          schema:
            type: string
        - name: type
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Equipment listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EquipmentListing"
    post:
      operationId: createEquipmentListing
      tags: [equipment]
      summary: List equipment for rent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEquipmentListingInput"
      responses:
        "201":
          description: Equipment listed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EquipmentListing"
  /equipment/bookings:
    post:
      operationId: bookEquipment
      tags: [equipment]
      summary: Book equipment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookEquipmentInput"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EquipmentBooking"
    get:
      operationId: getEquipmentBookings
      tags: [equipment]
      summary: Get equipment bookings
      responses:
        "200":
          description: Equipment bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EquipmentBooking"
  /transport/requests:
    get:
      operationId: getTransportRequests
      tags: [transport]
      summary: Get transport requests
      responses:
        "200":
          description: Transport requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TransportRequest"
    post:
      operationId: createTransportRequest
      tags: [transport]
      summary: Request produce transport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransportRequestInput"
      responses:
        "201":
          description: Transport request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransportRequest"
  /transport/requests/{id}:
    patch:
      operationId: updateTransportRequest
      tags: [transport]
      summary: Update transport request status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTransportStatusInput"
      responses:
        "200":
          description: Updated transport request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransportRequest"
...
components:
  schemas:
...
    EquipmentListing:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
        description:
          type: string
        ratePerDay:
          type: number
        location:
          type: string
        ownerName:
          type: string
        ownerPhone:
          type: string
        available:
          type: boolean
        availableFrom:
          type: string
        availableTo:
          type: string
        condition:
          type: string
          enum: [excellent, good, fair, poor]
        brand:
          type: string
        model:
          type: string
        yearOfMake:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - type
        - ratePerDay
        - location
        - ownerName
        - available
    CreateEquipmentListingInput:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        description:
          type: string
        ratePerDay:
          type: number
        location:
          type: string
        ownerName:
          type: string
        ownerPhone:
          type: string
        availableFrom:
          type: string
        availableTo:
          type: string
        condition:
          type: string
          enum: [excellent, good, fair, poor]
        brand:
          type: string
        model:
          type: string
        yearOfMake:
          type: integer
      required:
        - name
        - type
        - ratePerDay
        - location
        - ownerName
    EquipmentBooking:
      type: object
      properties:
        id:
          type: integer
        equipmentId:
          type: integer
        equipmentName:
          type: string
        fromDate:
          type: string
        toDate:
          type: string
        totalDays:
          type: integer
        totalCost:
          type: number
        farmerName:
          type: string
        farmerPhone:
          type: string
        status:
          type: string
          enum: [pending, confirmed, active, completed, cancelled]
        notes:
          type: string
        createdAt:
          type: string
      required:
        - id
        - equipmentId
        - equipmentName
        - fromDate
        - toDate
        - status
    BookEquipmentInput:
      type: object
      properties:
        equipmentId:
          type: integer
        fromDate:
          type: string
        toDate:
          type: string
        farmerName:
          type: string
        farmerPhone:
          type: string
        notes:
          type: string
      required:
        - equipmentId
        - fromDate
        - toDate
        - farmerName
    TransportRequest:
      type: object
      properties:
        id:
          type: integer
        farmerName:
          type: string
        farmerPhone:
          type: string
        pickupLocation:
          type: string
        dropLocation:
          type: string
        cropType:
          type: string
        quantity:
          type: number
        unit:
          type: string
        scheduledDate:
          type: string
        scheduledTime:
          type: string
        status:
          type: string
          enum: [pending, accepted, in_transit, delivered, cancelled]
        driverName:
          type: string
        driverPhone:
          type: string
        estimatedCost:
          type: number
        notes:
          type: string
        createdAt:
          type: string
      required:
        - id
        - farmerName
        - pickupLocation
        - dropLocation
        - cropType
        - quantity
        - scheduledDate
        - status
    CreateTransportRequestInput:
      type: object
      properties:
        farmerName:
          type: string
        farmerPhone:
          type: string
        pickupLocation:
          type: string
        dropLocation:
          type: string
        cropType:
          type: string
        quantity:
          type: number
        unit:
          type: string
        scheduledDate:
          type: string
        scheduledTime:
          type: string
        notes:
          type: string
      required:
        - farmerName
        - pickupLocation
        - dropLocation
        - cropType
        - quantity
        - scheduledDate
    UpdateTransportStatusInput:
      type: object
      properties:
        status:
          type: string
          enum: [pending, accepted, in_transit, delivered, cancelled]
        driverName:
          type: string
        driverPhone:
          type: string
        estimatedCost:
          type: number
      required:
        - status
...