# lib/api-zod/src/generated/types/createTransportRequestInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AgriSmart Agriculture Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateTransportRequestInput {
  farmerName: string;
  farmerPhone?: string;
  pickupLocation: string;
  dropLocation: string;
  cropType: string;
  quantity: number;
  unit?: string;
  scheduledDate: string;
  scheduledTime?: string;
  notes?: string;
}