# lib/api-zod/src/generated/types/equipmentBooking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AgriSmart Agriculture Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { EquipmentBookingStatus } from "./equipmentBookingStatus";

export interface EquipmentBooking {
  id: number;
  equipmentId: number;
  equipmentName: string;
  fromDate: string;
  toDate: string;
  totalDays?: number;
  totalCost?: number;
  farmerName?: string;
  farmerPhone?: string;
  status: EquipmentBookingStatus;
  notes?: string;
  createdAt?: string;
}