# lib/api-zod/src/generated/types/transportRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AgriSmart Agriculture Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { TransportRequestStatus } from "./transportRequestStatus";

export interface TransportRequest {
  id: number;
  farmerName: string;
  farmerPhone?: string;
  pickupLocation: string;
  dropLocation: string;
  cropType: string;
  quantity: number;
  unit?: string;
  scheduledDate: string;
  scheduledTime?: string;
  status: TransportRequestStatus;
  driverName?: string;
  driverPhone?: string;
  estimatedCost?: number;
  notes?: string;
  createdAt?: string;
}